add auth logic

+++ Core/Services/AdService.cs
@@ -16,7 +16,7 @@ using Sieve.Services;
 
 namespace Core.Services
 {
-    public class AdService : EntityService<Ad>
+    public class AdService : EntityService<Ad>, IAdService
     {
         private readonly IOptions<UserOptions> _userOptions;
         private readonly IImageHelper _imageHelper;
@@ -24,6 +24,12 @@ namespace Core.Services
         protected override void AdaptResponse<TResponse>(TResponse response)
         {
             if (response is IHasImage hasImage) _imageHelper.ImageNameToImageUrl(hasImage);
+        }
+
+        protected override bool IsAuthorized(Guid id, Guid userId)
+        {
+            if (id == Guid.Empty || userId == Guid.Empty) return false;
+            return Entries.SingleOrDefault(x => x.Id == id)?.UserId == userId;
         }
 
         public AdService(AdPortalContext context, IMapper mapper, ISieveProcessor sieveProcessor,
@@ -34,21 +40,23 @@ namespace Core.Services
             _imageHelper = imageHelper ?? throw new ArgumentNullException(nameof(imageHelper));
         }
 
-        public override async Task<TResponse> CreateNewAsync<TResponse>(IRequest request)
+        public async Task<TResponse> PostNewAdAsync<TResponse>(IRequest request, Guid userId)
+            where TResponse : IResponse
         {
             if (request == null) throw new ArgumentNullException(nameof(request));
-            if (!(request is AdCreateRequest adRequest)) throw new InvalidCastException(nameof(adRequest)); 
-            
+            if (!(request is AdCreateRequest adRequest)) throw new InvalidCastException(nameof(adRequest));
+
             await using var transaction = await Context.Database.BeginTransactionAsync(IsolationLevel.RepeatableRead);
 
-            var user = await Context.Users.SingleAsync(x => x.Id == adRequest.UserId);
+            var user = await Context.Users.SingleAsync(x => x.Id == userId);
             if (user.AdsCount >= _userOptions.Value.AdCountLimit)
-                throw new ConstraintException($"User {user.Id} has reached his ad limit");
+                throw new ConstraintException($"User {userId} has reached his ad limit");
 
             var ad = MapFromRequest(adRequest);
             ad.CreationDate = DateTime.Now;
             ad.ImageName = await _imageHelper.UploadImageAndGetNameAsync(adRequest.Image);
-
+            ad.UserId = userId;
+            
             Context.Ads.Add(ad);
             user.AdsCount++;
 
@@ -58,11 +66,12 @@ namespace Core.Services
             return MapToResponse<TResponse>(ad);
         }
 
-        public override async Task<TResponse> UpdateAsync<TResponse>(Guid id, IRequest request)
+        public override async Task<TResponse> UpdateAsync<TResponse>(Guid id, IRequest request, Guid userId)
         {
             if (request == null) throw new ArgumentNullException(nameof(request));
-            if (!(request is AdUpdateRequest adRequest)) throw new InvalidCastException(nameof(adRequest));  
-            
+            if (!(request is AdUpdateRequest adRequest)) throw new InvalidCastException(nameof(adRequest));
+            if (!IsAuthorized(id, userId)) throw new UnauthorizedAccessException();
+
             var ad = await Entries.SingleAsync(x => x.Id == id);
 
             _imageHelper.DeleteImage(ad.ImageName);
@@ -74,8 +83,10 @@ namespace Core.Services
             return MapToResponse<TResponse>(ad);
         }
 
-        public override async Task DeleteByIdAsync(Guid id)
+        public override async Task DeleteByIdAsync(Guid id, Guid userId)
         {
+            if (!IsAuthorized(id, userId)) throw new UnauthorizedAccessException();
+
             var ad = await Context.Ads.SingleAsync(x => x.Id == id);
 
             await using var transaction = await Context.Database.BeginTransactionAsync();
